pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "relepaper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Albert M", email = "rinkorn.alb@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "duckduckgo-search>=8.0.2",
    "fake-headers>=1.0.2",
    "fake-useragent>=2.2.0",
    "gradio>=5.31.0",
    "langchain>=0.3.23",
    "langchain-community>=0.3.21",
    "langchain-experimental>=0.3.4",
    "langchain-gigachat>=0.3.10",
    "langchain-ollama>=0.3.2",
    "langchain-openai>=0.3.14",
    "langgraph>=0.4.3",
    "matplotlib>=3.10.1",
    "numexpr>=2.10.2",
    "openreview-py>=1.47.2",
    "pyalex>=0.18",
    "selenium>=4.31.0",
    "selenium-stealth>=1.0.6",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "jupyter>=1.1.1",
    "pre-commit>=4.2.0",
    "ruff>=0.11.2",
    "transformers>=4.52.3",
]
lint = [
    "ruff>=0.11.2",
]
search = [
    "openalexapi>=0.0.1a0",
    "scholarly[tor]>=1.7.11",
]
test = [
    "pytest>=8.3.5",
    "pytest-cov>=6.1.1",
]

[tool.ruff]
line-length = 120

[tool.ruff.format]
docstring-code-line-length = 120
